userp/requirements.txt
alembic==1.5.8
bcrypt==3.2.0
decorator==5.0.9
elasticsearch==7.9.0
Flask==2.0.1
Flask-API==3.0.post1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
glog==0.3.1
numpy==1.18.2
oauthlib==3.1.0
PyJWT==2.1.0
psycopg2==2.9
requests==2.25.0
requests-oauthlib==1.3.0
SQLAlchemy==1.3.20